backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
lxml>=4.9.0
